GSM based SMS Application
This window application is created to send Multiple SMS using an external device called, GSM Modem. The GSM modem will be connected to the computer at the USB port. GSM Modem works on AT commands. This application will also read the messages  received in the SIM card of GSM Modem.